refactor(watch): unify watch integration in single WatchBridge component
WatchSessionSync and WatchRemoteInvalidationListener were both
side-effect-only, both guarded on Platform.OS === "ios", both returned
null, and mounted as adjacent siblings in _layout. One mount point makes
the watch integration easier to find and reason about. No behavior
change — the two useEffects are preserved as-is.
