+++ apps/mobile/app/_layout.tsx
@@ -19,8 +19,7 @@ import { SyncStatusBar } from "@/components/sync";
 import { TutorialOverlay } from "@/components/tutorial";
 import { GluestackUIProvider } from "@/components/ui";
 import { UpdatePrompt } from "@/components/update/UpdatePrompt";
-import { WatchRemoteInvalidationListener } from "@/components/WatchRemoteInvalidationListener";
-import { WatchSessionSync } from "@/components/WatchSessionSync";
+import { WatchBridge } from "@/components/WatchBridge";
 import { useAppUpdate } from "@/hooks/useAppUpdate";
 import { useSentryUserContext } from "@/hooks/useSentryUserContext";
 import { useStoreUpdate } from "@/hooks/useStoreUpdate";
@@ -320,8 +319,7 @@ export default function RootLayout() {
                                   <NavigationGuard>
                                     <BackgroundSyncHandler />
                                     <SentryUserContextHandler />
-                                    <WatchSessionSync />
-                                    <WatchRemoteInvalidationListener />
+                                    <WatchBridge />
                                     <NovuAutoSubscriber />
                                     <NotificationPromptHandler />
                                     <UpdatePromptHandler />
